Переименова контроллеры админки. Все предствления админки находятся по пути Views/Admin/... . Создан сервис AdminViewLocationExpander для указание пути до представление админки
