+++ Controllers/Admin/AdminAnswersController.cs
@@ -2,28 +2,27 @@
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using YouCan.Models;
-using static System.Net.Mime.MediaTypeNames;
 
 namespace YouCan.Controllers.Admin;
 
 [Route("Admin/[controller]/{action=index}")]
-public class QuestionsController : Controller
+public class AdminAnswersController : Controller
 {
     private readonly YouCanContext _context;
 
-    public QuestionsController(YouCanContext context)
+    public AdminAnswersController(YouCanContext context)
     {
         _context = context;
     }
 
-    // GET: Questions
+    // GET: Answers
     public async Task<IActionResult> Index()
     {
-        var questions = _context.Questions.Include(q => q.Test).Include(q => q.User);
-        return View(await questions.ToListAsync());
+        var answers = _context.Answers.Include(a => a.Question);
+        return View(await answers.ToListAsync());
     }
 
-    // GET: Questions/Details/5
+    // GET: Answers/Details/5
     public async Task<IActionResult> Details(int? id)
     {
         if (id == null)
@@ -31,59 +30,55 @@ public class QuestionsController : Controller
             return NotFound();
         }
 
-        var question = await _context.Questions
-            .Include(q => q.Test)
-            .Include(q => q.User)
-            .Include(q => q.Answers)
+        var answer = await _context.Answers
+            .Include(a => a.Question)
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (question == null)
+        if (answer == null)
         {
             return NotFound();
         }
 
-        return View(question);
+        return View(answer);
     }
 
-    // GET: Questions/Create
+    // GET: Answers/Create
     public async Task<IActionResult> Create(int id, string? returnUrl)
     {
-        var test = await _context.Tests.FindAsync(id);
-        if (test != null)
+        var question = await _context.Questions.FindAsync(id);
+        if (question != null)
         {
-            ViewBag.TestId = test.Id;
+            ViewBag.QuestionId = question.Id;    
             ViewBag.ReturnUrl = returnUrl;
-            return View();
         }
-        return NotFound();
+        return View();
     }
 
-    // POST: Questions/Create
+    // POST: Answers/Create
     // To protect from overposting attacks, enable the specific properties you want to bind to.
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Create(int id, Question question, string? returnUrl)
+    public async Task<IActionResult> Create(int id, Answer answer, string? returnUrl)
     {
         if (ModelState.IsValid)
         {
-            _context.Add(question);
+            _context.Add(answer);
             await _context.SaveChangesAsync();
             if (returnUrl != null && Url.IsLocalUrl(returnUrl))
             {
                 return Redirect(returnUrl);
             }
-            return RedirectToAction("Index");
+            return RedirectToAction(nameof(Index));
         }
-        var test = await _context.Tests.FindAsync(id);
-        if (test != null)
+        var question = await _context.Questions.FindAsync(id);
+        if (question != null)
         {
-            ViewBag.TestId = test.Id;
-            return View(question);
+            ViewBag.QuestionId = question.Id;
         }
-        return NotFound();
+        return View(answer);
     }
 
-    // GET: Questions/Edit/5
+    // GET: Answers/Edit/5
     public async Task<IActionResult> Edit(int? id)
     {
         if (id == null)
@@ -91,24 +86,23 @@ public class QuestionsController : Controller
             return NotFound();
         }
 
-        var question = await _context.Questions.FindAsync(id);
-        if (question == null)
+        var answer = await _context.Answers.FindAsync(id);
+        if (answer == null)
         {
             return NotFound();
         }
-        ViewData["TestId"] = new SelectList(_context.Tests, "Id", "Id", question.TestId);
-        ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", question.UserId);
-        return View(question);
+        ViewData["QuestionId"] = new SelectList(_context.Questions, "Id", "Id", answer.QuestionId);
+        return View(answer);
     }
 
-    // POST: Questions/Edit/5
+    // POST: Answers/Edit/5
     // To protect from overposting attacks, enable the specific properties you want to bind to.
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Edit(int id, [Bind("Id,Title,Content,Type,IsPublished,TestId,UserId")] Question question)
+    public async Task<IActionResult> Edit(int id, [Bind("Id,Text,IsCorrect,QuestionId")] Answer answer)
     {
-        if (id != question.Id)
+        if (id != answer.Id)
         {
             return NotFound();
         }
@@ -117,12 +111,12 @@ public class QuestionsController : Controller
         {
             try
             {
-                _context.Update(question);
+                _context.Update(answer);
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!QuestionExists(question.Id))
+                if (!AnswerExists(answer.Id))
                 {
                     return NotFound();
                 }
@@ -133,12 +127,11 @@ public class QuestionsController : Controller
             }
             return RedirectToAction(nameof(Index));
         }
-        ViewData["TestId"] = new SelectList(_context.Tests, "Id", "Id", question.TestId);
-        ViewData["UserId"] = new SelectList(_context.Users, "Id", "Id", question.UserId);
-        return View(question);
+        ViewData["QuestionId"] = new SelectList(_context.Questions, "Id", "Id", answer.QuestionId);
+        return View(answer);
     }
 
-    // GET: Questions/Delete/5
+    // GET: Answers/Delete/5
     public async Task<IActionResult> Delete(int? id)
     {
         if (id == null)
@@ -146,35 +139,34 @@ public class QuestionsController : Controller
             return NotFound();
         }
 
-        var question = await _context.Questions
-            .Include(q => q.Test)
-            .Include(q => q.User)
+        var answer = await _context.Answers
+            .Include(a => a.Question)
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (question == null)
+        if (answer == null)
         {
             return NotFound();
         }
 
-        return View(question);
+        return View(answer);
     }
 
-    // POST: Questions/Delete/5
+    // POST: Answers/Delete/5
     [HttpPost, ActionName("Delete")]
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
-        var question = await _context.Questions.FindAsync(id);
-        if (question != null)
+        var answer = await _context.Answers.FindAsync(id);
+        if (answer != null)
         {
-            _context.Questions.Remove(question);
+            _context.Answers.Remove(answer);
         }
 
         await _context.SaveChangesAsync();
         return RedirectToAction(nameof(Index));
     }
 
-    private bool QuestionExists(int id)
+    private bool AnswerExists(int id)
     {
-        return _context.Questions.Any(e => e.Id == id);
+        return _context.Answers.Any(e => e.Id == id);
     }
 }
