+++ Controllers/Admin/AdminTopicsController.cs
@@ -1,28 +1,25 @@
 ﻿using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using YouCan.Models;
 
 namespace YouCan.Controllers.Admin;
 
-[Route("Admin/[controller]/{action=index}")]
-public class AnswersController : Controller
+public class AdminTopicsController : Controller
 {
     private readonly YouCanContext _context;
 
-    public AnswersController(YouCanContext context)
+    public AdminTopicsController(YouCanContext context)
     {
         _context = context;
     }
 
-    // GET: Answers
+    // GET: Topics
     public async Task<IActionResult> Index()
     {
-        var answers = _context.Answers.Include(a => a.Question);
-        return View(await answers.ToListAsync());
+        return View(await _context.Topics.ToListAsync());
     }
 
-    // GET: Answers/Details/5
+    // GET: Topics/Details/5
     public async Task<IActionResult> Details(int? id)
     {
         if (id == null)
@@ -30,55 +27,40 @@ public class AnswersController : Controller
             return NotFound();
         }
 
-        var answer = await _context.Answers
-            .Include(a => a.Question)
+        var topic = await _context.Topics
+            .Include(t => t.Subtopics)
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (answer == null)
+        if (topic == null)
         {
             return NotFound();
         }
 
-        return View(answer);
+        return View(topic);
     }
 
-    // GET: Answers/Create
-    public async Task<IActionResult> Create(int id, string? returnUrl)
+    // GET: Topics/Create
+    public IActionResult Create()
     {
-        var question = await _context.Questions.FindAsync(id);
-        if (question != null)
-        {
-            ViewBag.QuestionId = question.Id;    
-            ViewBag.ReturnUrl = returnUrl;
-        }
         return View();
     }
 
-    // POST: Answers/Create
+    // POST: Topics/Create
     // To protect from overposting attacks, enable the specific properties you want to bind to.
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Create(int id, Answer answer, string? returnUrl)
+    public async Task<IActionResult> Create(Topic topic)
     {
         if (ModelState.IsValid)
         {
-            _context.Add(answer);
+            _context.Add(topic);
             await _context.SaveChangesAsync();
-            if (returnUrl != null && Url.IsLocalUrl(returnUrl))
-            {
-                return Redirect(returnUrl);
-            }
             return RedirectToAction(nameof(Index));
         }
-        var question = await _context.Questions.FindAsync(id);
-        if (question != null)
-        {
-            ViewBag.QuestionId = question.Id;
-        }
-        return View(answer);
+        return View(topic);
     }
 
-    // GET: Answers/Edit/5
+    // GET: Topics/Edit/5
     public async Task<IActionResult> Edit(int? id)
     {
         if (id == null)
@@ -86,23 +68,22 @@ public class AnswersController : Controller
             return NotFound();
         }
 
-        var answer = await _context.Answers.FindAsync(id);
-        if (answer == null)
+        var topic = await _context.Topics.FindAsync(id);
+        if (topic == null)
         {
             return NotFound();
         }
-        ViewData["QuestionId"] = new SelectList(_context.Questions, "Id", "Id", answer.QuestionId);
-        return View(answer);
+        return View(topic);
     }
 
-    // POST: Answers/Edit/5
+    // POST: Topics/Edit/5
     // To protect from overposting attacks, enable the specific properties you want to bind to.
     // For more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
     [HttpPost]
     [ValidateAntiForgeryToken]
-    public async Task<IActionResult> Edit(int id, [Bind("Id,Text,IsCorrect,QuestionId")] Answer answer)
+    public async Task<IActionResult> Edit(int id, Topic topic)
     {
-        if (id != answer.Id)
+        if (id != topic.Id)
         {
             return NotFound();
         }
@@ -111,12 +92,12 @@ public class AnswersController : Controller
         {
             try
             {
-                _context.Update(answer);
+                _context.Update(topic);
                 await _context.SaveChangesAsync();
             }
             catch (DbUpdateConcurrencyException)
             {
-                if (!AnswerExists(answer.Id))
+                if (!TopicExists(topic.Id))
                 {
                     return NotFound();
                 }
@@ -127,11 +108,10 @@ public class AnswersController : Controller
             }
             return RedirectToAction(nameof(Index));
         }
-        ViewData["QuestionId"] = new SelectList(_context.Questions, "Id", "Id", answer.QuestionId);
-        return View(answer);
+        return View(topic);
     }
 
-    // GET: Answers/Delete/5
+    // GET: Topics/Delete/5
     public async Task<IActionResult> Delete(int? id)
     {
         if (id == null)
@@ -139,34 +119,33 @@ public class AnswersController : Controller
             return NotFound();
         }
 
-        var answer = await _context.Answers
-            .Include(a => a.Question)
+        var topic = await _context.Topics
             .FirstOrDefaultAsync(m => m.Id == id);
-        if (answer == null)
+        if (topic == null)
         {
             return NotFound();
         }
 
-        return View(answer);
+        return View(topic);
     }
 
-    // POST: Answers/Delete/5
+    // POST: Topics/Delete/5
     [HttpPost, ActionName("Delete")]
     [ValidateAntiForgeryToken]
     public async Task<IActionResult> DeleteConfirmed(int id)
     {
-        var answer = await _context.Answers.FindAsync(id);
-        if (answer != null)
+        var topic = await _context.Topics.FindAsync(id);
+        if (topic != null)
         {
-            _context.Answers.Remove(answer);
+            _context.Topics.Remove(topic);
         }
 
         await _context.SaveChangesAsync();
         return RedirectToAction(nameof(Index));
     }
 
-    private bool AnswerExists(int id)
+    private bool TopicExists(int id)
     {
-        return _context.Answers.Any(e => e.Id == id);
+        return _context.Topics.Any(e => e.Id == id);
     }
 }
