+++ Views/Admin/AdminTopics/Delete.cshtml
